requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
exceptiongroup==1.2.2
    # via anyio
flask==3.1.0
    # via -r requirements.in
google-api-core[grpc]==2.24.2
    # via
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-translate
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-translate
    #   google-genai
google-cloud-core==2.4.3
    # via
    #   google-cloud-firestore
    #   google-cloud-translate
google-cloud-firestore==2.20.1
    # via -r requirements.in
google-cloud-translate==3.20.2
    # via -r requirements.in
google-genai==1.7.0
    # via -r requirements.in
googleapis-common-protos[grpc]==1.69.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-translate
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-cloud-translate
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   google-cloud-translate
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via google-genai
pydantic-core==2.27.2
    # via pydantic
requests==2.32.3
    # via
    #   google-api-core
    #   google-genai
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via flask
